First Pass
Fixed Jittering by using FixedUpdate
Renamed Velocity Indicator to Camera Target
Used Transform.Translate() instead of updating
raw positions.
